requirements.txt
nltk
ipykernel
numpy
pandas
textblob
requests
python-dotenv
matplotlib